+++ src/assets/Icons.js
@@ -67,4 +67,11 @@ export const LeftDownIcon = () => {
     )
 }
 
+export const HomeIcon = () => {
+    return (
+        <svg xmlns="http://www.w3.org/2000/svg" height="48" viewBox="0 -960 960 960" width="48" fill="#777777"><path d="M240-200h120v-240h240v240h120v-360L480-740 240-560v360Zm-80 80v-480l320-240 320 240v480H520v-240h-80v240H160Zm320-350Z"/></svg>
+    )
+}
+
+
